+++ src/components/interview/Interview/Interview.css
@@ -1,0 +1,56 @@
+.shadow {box-shadow: 0 16px 24px 2px rgba(0,0,0,0.15), 0 6px 30px 5px rgba(0,0,0,0.12), 0 8px 10px -7px rgba(0,0,0,0.2);}
+
+.box-fix {
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+}
+
+.question-container {
+  position: fixed;
+  width: 100vw;
+  height: calc(100vh - 3rem);
+  transition: transform 0.5s linear;
+}
+
+.question-container-post {
+  transform: translate(-100vw, 0);
+}
+
+.deck-btn-panel {
+  background: #48a5a0;
+  border-radius: 0.5rem;
+  padding: 0.5rem 1rem;
+
+  position: fixed;
+  bottom: 2rem;
+  top: calc(100% - 5rem);
+  left: calc(100% - 30% + 4rem);
+}
+@media only screen and (max-width: 800px) {.deck-btn-panel {display: none;}}
+@media only screen and (min-width: 800px) {.deck-btn-panel {left: calc(100% - 20% + 4rem);}}
+@media only screen and (min-width: 1400px) {.deck-btn-panel {left: calc(100% - 30% + 4rem);}}
+
+.back-btn {
+  height: 2rem; width: 2rem;
+  background: #f03750;
+  border-radius: 1rem;
+  border: 0;
+  color: #9b0528;
+  font-size: 1.5rem;
+  cursor: pointer;
+  margin-right: 1rem;
+}
+.back-btn:focus {outline: 0;}
+
+.end-btn {
+  height: 2rem;
+  background: #faaf05;
+  border-radius: 0.2rem;
+  border: 0;
+  color: #965f0a;
+  font-size: 1.5rem;
+  cursor: pointer;
+}
+.end-btn:focus {outline: 0;}
+
