+++ src/components/interview/Interview/Interview.js
@@ -1,6 +1,7 @@
-import React, {useState, useContext} from 'react';
-import QuestionDeck from '../QuestionDeck/QuestionDeck';
+import React, {useState} from 'react';
+import {Swipeable} from 'react-swipeable';
 import InterviewEnd from '../InterviewEnd/InterviewEnd';
+import Question from '../Question/Question';
 import './Interview.css';
 import gql from 'graphql-tag';
 import {useQuery} from '@apollo/react-hooks';
@@ -25,6 +26,25 @@ export const Interview = props => {
   if (loading) return <p>Loading...</p>;
   if (error) return <p>Error...</p>;
 
+  const populateCards = () => {
+    return data.questions.map(({id, body}, index) => {
+      return <Question cur={cur} id={id} key={id} pos={index + 1} question={body} fs={{note: updateNote, score: updateScore, next: nextCard}}/>
+    });
+  }
+
+  const nextCard = id => {
+    setCur(cur + 1);
+  }
+
+  const prevCard = () => {
+    let nextCur = cur - 1;
+    if (nextCur < 1) {
+      setCur(1);
+    } else {
+      setCur(nextCur);
+    }
+  }
+
   const updateNote = (id, body) => {
     if (notes[id]) {
       notes[id] = {...notes[id], body: body};
@@ -35,9 +55,9 @@ export const Interview = props => {
 
   const updateScore = (id, score) => {
     if (notes[id]) {
-      notes[id] = {...notes[id], score: score};
+      notes[id] = {...notes[id], score: parseInt(score)};
     } else {
-      notes[id] = {body: '', score: score};
+      notes[id] = {body: '', score: parseInt(score)};
     }
   }
 
@@ -50,9 +70,22 @@ export const Interview = props => {
     setFocus(0);
   }
 
+  const isFocused = () => {
+    if (focus === 0) return ''
+    return ' question-container-post';
+  }
+
   return (
     <div>
-      <QuestionDeck data={data} fs={{note: updateNote, score: updateScore, end: endInterview}} focused={focus === 0}/>
+      <div className={'question-container' + isFocused()}>
+        <Swipeable onSwipedLeft={event => nextCard()} onSwipedRight={event => prevCard()} trackMouse={true} preventDefaultTouchmoveEvent={true}>
+          {populateCards()}
+        </Swipeable>
+        <div className='deck-btn-panel shadow'>
+          <button className='back-btn' onClick={event => prevCard()}>🠘</button>
+          <button className='end-btn' onClick={event => endInterview()}>End</button>
+        </div>
+      </div>
       <InterviewEnd fs={{note: updateNote, score: updateScore}} focused={focus === 1}/>
       <button style={{position: 'fixed', top: '4rem'}} onClick={neverMind}>Nope</button>
     </div>
